fullstack/backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
celery==5.3.4
redis==5.0.1
Pillow>=10.2.0
reportlab==4.0.7
qrcode==7.4.2
openpyxl==3.1.2
python-decouple==3.8
django-extensions==3.2.3
whitenoise==6.6.0
gunicorn==21.2.0
django-filter
pandas
playwright
